requirements.txt
Django>=5.1,<6.1
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
mysqlclient>=2.2